Added even more functionality

+++ dataservice/Models/Trainingsbook.cs
@@ -5,7 +5,7 @@ namespace dataservice.Models
     public partial class Trainingsbook
     {
         public int TrainingId { get; set; }
-        public int Isbn { get; set; }
+        public long Isbn { get; set; }
         [JsonIgnore]
         public Book IsbnNavigation { get; set; }
         [JsonIgnore]
